util/jobqueue.py: fix gobject.source_remove warning spam
When enabling debug output, we're currently seeing this sequence GTimer
instance until the end :

> Queueing CycleTimer on default for execution in 10.000000
> Queued CycleTimer
> ...
> Executing GTimer CycleTimer
> ...
> Executed GTimer CycleTimer
> ...
> Cancelled GTimer CycleTimer
> ...
> ...
> Queueing CycleTimer on default for execution in 10.000000
> [Another CycleTimer queuing-executing-cancelling sequence]

Currently, the callback function GTimer::exec_item() returns nothing. In a
boolean comparison, it is equivalent to return False. The documentation [1]
says that « The [callback] function is called repeatedly until it returns FALSE,
at which point the timeout is automatically destroyed and the function will not
be called again. ».

Here is the problem : when GTimer::cancel() is called, it unconditionnaly tries
to remove the event source using gobject.source_remove(). But, if the callback
function was already executed before the ::cancel() call, the timeout was
already destroyed, and source_remove() produces this warning :
> /usr/lib64/python2.7/site-packages/gnome15/util/jobqueue.py:92: Warning:
> Source ID 105 was not found when attempting to remove it
> gobject.source_remove(self.source)

This warning comes up twice every ten seconds or so, and leads to a
warning spam in regular output (usually ~/.xsession-errors).

When ride up to the glib documentation [2], the g_source_remove() doc clearly
stated that « It is a programmer error to attempt to remove a non-existent
source ».

This commit is doing two things :
 - it returns explicitly False from the callback function, to follow up the
   documentation and leave no doubt.
 - it checks if the callback function has run (and if the timeout was already
   destroyed) before calling gobject.source_remove(), avoiding the warning
   spam.

[1] http://www.pygtk.org/pygtk2reference/gobject-functions.html#function-gobject--timeout-add
[2] https://developer.gnome.org/glib/stable/glib-The-Main-Event-Loop.html#g-source-remove